chore(release): add git commit message for version tagging
This change adds a git commit message for version tagging in the release configuration of nx.json file. The commit message will be in the format "chore(release): 🔖 ${version}", where ${version} will be replaced with the actual version number. This will improve clarity and consistency when tagging versions in the repository.